requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.0.0
albumentations>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
Pillow>=10.0.0
lxml>=4.9.0

